# services/id/src/updspaceid/openapi.yaml
openapi: 3.1.0
info:
  title: UpdSpaceID
  version: 1.0.0
  description: |
    UpdSpaceID is a multi-tenant identity provider.

    Notes:
    - All requests MUST include `X-Request-Id`.
    - Tenant context MUST be provided by BFF via `X-Tenant-Id` and `X-Tenant-Slug`.
    - All errors use a unified envelope: `{ "error": { ... } }`.
servers:
  - url: https://id.updspace.com
    description: Production
  - url: http://localhost:8001
    description: Local dev
components:
  securitySchemes:
    BffSession:
      type: apiKey
      in: cookie
      name: updspace_session
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: true
      schema: { type: string }
    XTenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
    XTenantSlug:
      name: X-Tenant-Slug
      in: header
      required: true
      schema: { type: string }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
...
    MagicLinkRequestIn:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    MagicLinkConsumeIn:
      type: object
      required: [token]
      properties:
        token: { type: string }
    SessionIssueOut:
      type: object
      required: [ok, user_id, session_token]
      properties:
        ok: { type: boolean, const: true }
        user_id: { type: string, format: uuid }
        session_token: { type: string }
...
paths:
...
  /api/v1/auth/magic-link/request:
    post:
      operationId: auth_magic_link_request
      parameters: [{$ref: '#/components/parameters/XRequestId'}, {$ref: '#/components/parameters/XTenantId'}, {$ref: '#/components/parameters/XTenantSlug'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MagicLinkRequestIn' }
      responses:
        '200':
          description: OK
        '429':
          description: RATE_LIMITED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/v1/auth/magic-link/consume:
    post:
      operationId: auth_magic_link_consume
      parameters: [{$ref: '#/components/parameters/XRequestId'}, {$ref: '#/components/parameters/XTenantId'}, {$ref: '#/components/parameters/XTenantSlug'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MagicLinkConsumeIn' }
      responses:
        '200':
          description: Issues server-side session for BFF
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionIssueOut' }
        '403':
          description: ACCOUNT_SUSPENDED | ACCOUNT_BANNED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...